AP_FETtecOneWire: Temperature can be negative

+++ libraries/AP_FETtecOneWire/AP_FETtecOneWire.cpp
@@ -107,7 +107,7 @@ void AP_FETtecOneWire::update()
             if (mask & esc_mask) { // only update telemetry of enabled ESCs
                 switch(_telem_avail) {
                 case telem_type::TEMP:
-                    t.temperature_deg = requestedTelemetry[i],
+                    t.temperature_deg = int16_t(requestedTelemetry[i]),
                     update_telem_data(i, t, AP_ESC_Telem_Backend::TelemetryType::TEMPERATURE);
                     break;
